separate document uploads into own component
