+++ src/DashboardUI/src/pages/application/create/ApplicationCreatePage.tsx
@@ -1,18 +1,11 @@
-import { useMsal } from '@azure/msal-react';
-import { mdiTrashCanOutline } from '@mdi/js';
-import Icon from '@mdi/react';
 import { createRef, useRef, useState } from 'react';
-import Alert from 'react-bootstrap/esm/Alert';
 import Button from 'react-bootstrap/esm/Button';
 import Form from 'react-bootstrap/esm/Form';
 import InputGroup from 'react-bootstrap/esm/InputGroup';
-import { useMutation } from 'react-query';
 import { useNavigate } from 'react-router-dom';
-import { uploadApplicationDocuments } from '../../../accessors/applicationAccessor';
 import { NotImplementedPlaceholder } from '../../../components/NotImplementedAlert';
 import { states } from '../../../config/states';
 import { useConservationApplicationContext } from '../../../contexts/ConservationApplicationProvider';
-import { ApplicationDocument } from '../../../data-contracts/ApplicationDocuments';
 import { ApplicationSubmissionForm } from '../../../data-contracts/ApplicationSubmissionForm';
 import {
   CompensationRateUnits,
@@ -23,6 +16,7 @@ import {
 import { formatNumber } from '../../../utilities/valueFormatters';
 import ApplicationFormSection from '../components/ApplicationFormSection';
 import { ApplicationNavbar } from '../components/ApplicationNavbar';
+import { ApplicationDocumentUpload } from './ApplicationDocumentUpload';
 
 export function ApplicationCreatePage() {
   const { state } = useConservationApplicationContext();
@@ -52,14 +46,12 @@ const responsiveOneThirdWidthDefault = 'col-lg-4 col-md-6 col-12';
 const responsiveHalfWidthDefault = 'col-lg-6 col-12';
 
 function ApplicationCreatePageForm() {
-  const msalContext = useMsal();
   const navigate = useNavigate();
   const { state, dispatch } = useConservationApplicationContext();
   const stateForm = state.conservationApplication.applicationSubmissionForm;
   const polygonData = state.conservationApplication.estimateLocations;
 
   const [formValidated, setFormValidated] = useState(false);
-  const [uploadDocumentErrorMessage, setUploadDocumentErrorMessage] = useState<string | null>(null);
 
   const navigateToReviewApplicationPage = () => {
     navigate(`/application/${state.conservationApplication.waterConservationApplicationId}/review`);
@@ -158,89 +150,6 @@ function ApplicationCreatePageForm() {
     if (isFormValid) {
       navigateToReviewApplicationPage();
     }
-  };
-
-  const uploadDocumentMutation = useMutation({
-    mutationFn: async (params: { files: File[] }) => {
-      clearUploadDocumentError();
-      return await uploadApplicationDocuments(msalContext, params.files);
-    },
-    onSuccess: (uploadedDocuments: ApplicationDocument[]) => {
-      dispatch({
-        type: 'APPLICATION_DOCUMENT_UPLOADED',
-        payload: {
-          uploadedDocuments,
-        },
-      });
-    },
-    onError: () => {
-      setUploadDocumentError('An error occurred while uploading the document(s). Please try again.');
-    },
-  });
-
-  const MAX_NUMBER_UPLOADED_DOCUMENTS = 10;
-  const UPLOADED_DOCUMENT_MAX_SIZE_MB = 25;
-
-  const isOverMaxDocumentLimit = (files: File[]): boolean => {
-    return state.conservationApplication.supportingDocuments.length + files.length > MAX_NUMBER_UPLOADED_DOCUMENTS;
-  };
-
-  const isOverMaxSizeLimit = (files: File[]): boolean => {
-    return files.some((file) => file.size > UPLOADED_DOCUMENT_MAX_SIZE_MB * 1024 * 1024);
-  };
-
-  const handleUploadDocument = () => {
-    const fileInput = document.createElement('input');
-    fileInput.type = 'file';
-    fileInput.multiple = true;
-    fileInput.onchange = async (e) => {
-      const fileList = (e.target as HTMLInputElement).files;
-      if (fileList && fileList.length > 0) {
-        const files = Array.from(fileList);
-
-        if (isOverMaxDocumentLimit(files)) {
-          setUploadDocumentError('You can upload a total of 10 documents.');
-          return;
-        }
-
-        if (isOverMaxSizeLimit(files)) {
-          setUploadDocumentError('Each file must be smaller than 25MB.');
-          return;
-        }
-
-        uploadDocumentMutation.mutate({ files });
-      }
-    };
-    fileInput.click();
-  };
-
-  const setUploadDocumentError = (message: string) => {
-    setUploadDocumentErrorMessage(message);
-  };
-
-  const clearUploadDocumentError = () => {
-    setUploadDocumentErrorMessage(null);
-  };
-
-  const handleRemoveDocument = (blobName: string) => {
-    dispatch({
-      type: 'APPLICATION_DOCUMENT_REMOVED',
-      payload: {
-        removedBlobName: blobName,
-      },
-    });
-  };
-
-  // TODO: JN - validation for form for upload docs
-
-  const handleDocumentDescriptionChanged = (blobName: string, description: string) => {
-    dispatch({
-      type: 'APPLICATION_DOCUMENT_UPDATED',
-      payload: {
-        blobName,
-        description,
-      },
-    });
   };
 
   return (
@@ -652,56 +561,12 @@ function ApplicationCreatePageForm() {
       </Form>
 
       <ApplicationFormSection title="Supporting Documents (Optional)">
-        {uploadDocumentErrorMessage !== null && (
-          <Alert variant="danger" dismissible onClose={clearUploadDocumentError}>
-            {uploadDocumentErrorMessage}
-          </Alert>
-        )}
-        <div className="col mb-4">
-          {state.conservationApplication.supportingDocuments.length > 0 && (
-            <table className="table">
-              <tbody>
-                {state.conservationApplication.supportingDocuments.map((file, index) => (
-                  <tr key={`${file.fileName}-${index}`}>
-                    <td className="align-content-center">{file.fileName}</td>
-                    <td className="align-content-center flex-grow-1">
-                      <Form.Group controlId={`supportingDocumentsDescription-${index}`}>
-                        <Form.Control
-                          as="textarea"
-                          maxLength={4000}
-                          value={file.description}
-                          onChange={(e) => handleDocumentDescriptionChanged(file.blobName, e.target.value)}
-                        ></Form.Control>
-                      </Form.Group>
-                    </td>
-                    <td className="align-content-center text-center">
-                      <Button
-                        variant="link"
-                        className="px-1 py-1 text-danger"
-                        onClick={() => handleRemoveDocument(file.blobName)}
-                      >
-                        <Icon
-                          path={mdiTrashCanOutline}
-                          size="1.5em"
-                          aria-label="Remove supporting document button"
-                        ></Icon>
-                      </Button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          )}
-          <Button variant="outline-primary" onClick={handleUploadDocument}>
-            Upload
-          </Button>
-        </div>
+        <ApplicationDocumentUpload></ApplicationDocumentUpload>
       </ApplicationFormSection>
 
       <hr className="m-0" />
       <div className="d-flex justify-content-end p-3">
         <Button variant="success" onClick={handleSubmitClicked}>
-          {/* TODO: JN - onClick={handleSubmit} */}
           Review & Submit
         </Button>
       </div>
